Add default EC2 IAM policies when scheduler is byos
