+++ cli/src/pcluster/templates/slurm_builder.py
@@ -19,7 +19,7 @@ from aws_cdk.core import CfnCustomResource, CfnDeletionPolicy, CfnOutput, CfnPar
 
 from pcluster.aws.aws_api import AWSApi
 from pcluster.config.cluster_config import SlurmClusterConfig
-from pcluster.constants import IAM_ROLE_PATH, PCLUSTER_CLUSTER_NAME_TAG, PCLUSTER_DYNAMODB_PREFIX
+from pcluster.constants import PCLUSTER_DYNAMODB_PREFIX
 from pcluster.models.s3_bucket import S3Bucket
 from pcluster.templates.cdk_builder_utils import (
     PclusterLambdaConstruct,
@@ -74,10 +74,6 @@ class SlurmConstruct(Construct):
 
     def _format_arn(self, **kwargs):
         return Stack.of(self).format_arn(**kwargs)
-
-    def _cluster_scoped_iam_path(self):
-        """Return a path to be associated IAM roles and instance profiles."""
-        return f"{IAM_ROLE_PATH}{self.stack_name}/"
 
     # -- Parameters -------------------------------------------------------------------------------------------------- #
 
@@ -157,41 +153,6 @@ class SlurmConstruct(Construct):
                 "resources": [
                     self._format_arn(service="dynamodb", resource=f"table/{PCLUSTER_DYNAMODB_PREFIX}{self.stack_name}")
                 ],
-            },
-            {
-                "sid": "EC2Terminate",
-                "effect": iam.Effect.ALLOW,
-                "actions": ["ec2:TerminateInstances"],
-                "resources": ["*"],
-                "conditions": {"StringEquals": {f"ec2:ResourceTag/{PCLUSTER_CLUSTER_NAME_TAG}": self.stack_name}},
-            },
-            {
-                "sid": "EC2RunInstances",
-                "effect": iam.Effect.ALLOW,
-                "actions": ["ec2:RunInstances"],
-                "resources": [
-                    self._format_arn(service="ec2", resource=f"subnet/{subnet_id}")
-                    for subnet_id in self.config.compute_subnet_ids
-                ]
-                + [
-                    self._format_arn(service="ec2", resource="network-interface/*"),
-                    self._format_arn(service="ec2", resource="instance/*"),
-                    self._format_arn(service="ec2", resource="volume/*"),
-                    self._format_arn(service="ec2", resource=f"key-pair/{self.config.head_node.ssh.key_name}"),
-                    self._format_arn(service="ec2", resource="security-group/*"),
-                    self._format_arn(service="ec2", resource="launch-template/*"),
-                    self._format_arn(service="ec2", resource="placement-group/*"),
-                ]
-                + [
-                    self._format_arn(service="ec2", resource=f"image/{queue_ami}", account="")
-                    for _, queue_ami in self.config.image_dict.items()
-                ],
-            },
-            {
-                "sid": "PassRole",
-                "actions": ["iam:PassRole"],
-                "effect": iam.Effect.ALLOW,
-                "resources": self._generate_head_node_pass_role_resources(),
             },
         ]
         policy_name = "parallelcluster-slurm-head-node"
@@ -321,33 +282,6 @@ class SlurmConstruct(Construct):
         )
 
         return cluster_hosted_zone
-
-    def _generate_head_node_pass_role_resources(self):
-        """Return a unique list of ARNs that the head node should be able to use when calling PassRole."""
-        default_pass_role_resource = self._format_arn(
-            service="iam",
-            region="",
-            resource=f"role{self._cluster_scoped_iam_path()}*",
-        )
-
-        # If there are any queues where a custom instance role was specified,
-        # enable the head node to pass permissions to those roles.
-        custom_queue_role_arns = {
-            arn for queue in self.config.scheduling.queues for arn in queue.iam.instance_role_arns
-        }
-        if custom_queue_role_arns:
-            pass_role_resources = custom_queue_role_arns
-
-            # Include the default IAM role path for the queues that
-            # aren't using a custom instance role.
-            queues_without_custom_roles = [
-                queue for queue in self.config.scheduling.queues if not queue.iam.instance_role_arns
-            ]
-            if any(queues_without_custom_roles):
-                pass_role_resources.add(default_pass_role_resource)
-        else:
-            pass_role_resources = {default_pass_role_resource}
-        return list(pass_role_resources)
 
     # -- Conditions -------------------------------------------------------------------------------------------------- #
